Add ddl test cases for multiple sharding strategy 1th.
