+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/AbstractShardingDatabaseOnlyTest.java
@@ -75,11 +75,19 @@ public abstract class AbstractShardingDatabaseOnlyTest extends AbstractSQLAssert
             final ShardingRuleConfiguration shardingRuleConfig = new ShardingRuleConfiguration();
             TableRuleConfiguration orderTableRuleConfig = new TableRuleConfiguration();
             orderTableRuleConfig.setLogicTable("t_order");
+            orderTableRuleConfig.setLogicIndex("t_order_index");
             orderTableRuleConfig.setKeyGeneratorColumnName("order_id");
             orderTableRuleConfig.setKeyGeneratorClass(IncrementKeyGenerator.class.getName());
             shardingRuleConfig.getTableRuleConfigs().add(orderTableRuleConfig);
             TableRuleConfiguration orderItemTableRuleConfig = new TableRuleConfiguration();
             orderItemTableRuleConfig.setLogicTable("t_order_item");
+            TableRuleConfiguration logTableRuleConfig = new TableRuleConfiguration();
+            logTableRuleConfig.setLogicIndex("t_log_index");
+            logTableRuleConfig.setLogicTable("t_log");
+            TableRuleConfiguration tempLogTableRuleConfig = new TableRuleConfiguration();
+            tempLogTableRuleConfig.setLogicTable("t_temp_log");
+            shardingRuleConfig.getTableRuleConfigs().add(logTableRuleConfig);
+            shardingRuleConfig.getTableRuleConfigs().add(tempLogTableRuleConfig);
             shardingRuleConfig.getTableRuleConfigs().add(orderItemTableRuleConfig);
             shardingRuleConfig.getBindingTableGroups().add("t_order, t_order_item");
             shardingRuleConfig.setDefaultDatabaseShardingStrategyConfig(new ComplexShardingStrategyConfiguration("user_id", ComplexKeysModuloDatabaseShardingAlgorithm.class.getName()));
