+++ sharding-jdbc-core/src/test/resources/integrate/dataset/masterslave/expect/drop/masterslaveonly.xml
@@ -1,0 +1,2 @@
+<dataset>
+</dataset>
