+++ sharding-jdbc-core/src/test/resources/integrate/dataset/masterslave/expect/index/masterslaveonly.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
