+++ sharding-jdbc-core/src/test/resources/integrate/dataset/masterslave/expect/index/masterslaveonly_sharding.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_order />
+</dataset>
