+++ sharding-jdbc-core/src/test/resources/integrate/dataset/masterslave/expect/truncate/masterslaveonly.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
