+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/db/expect/drop/db.xml
@@ -1,0 +1,2 @@
+<dataset>
+</dataset>
