+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/db/expect/index/db.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
