+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/db/expect/index/db_sharding.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_order />
+</dataset>
