+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/dbtbl/expect/drop/dbtbl.xml
@@ -1,0 +1,2 @@
+<dataset>
+</dataset>
