+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/dbtbl/expect/index/dbtbl.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
