+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/dbtbl/expect/truncate/dbtbl.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
