+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/expect/create/masterslave.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
