+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/expect/drop/masterslave.xml
@@ -1,0 +1,2 @@
+<dataset>
+</dataset>
