+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/expect/index/masterslave_sharding.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_order />
+</dataset>
