+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/expect/truncate/masterslave.xml
@@ -1,0 +1,3 @@
+<dataset>
+    <t_log />
+</dataset>
